well done

+++ src/main/java/org/example/DAO/UserDao.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserDao {
     List<User> getAll();
     void add(User user);
-    void delete(int id);
-    void update(int id, User user);
-    User findById(int id);
+    void delete(Long id);
+    void update(Long id, User user);
+    User findById(Long id);
 }
